@@ -1,13 +1,15 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
+import { environment } from 'src/environments/environment';
 import { Habilidades } from '../model/habilidades';
 
 @Injectable({
   providedIn: 'root'
 })
 export class HabilidadesService {
-  URL = 'http://localhost:8080/skill/';
+  //URL = 'http://localhost:8080/skill/';
+  URL = environment.EURL + '/skill/';
 
   constructor(private httpClient: HttpClient) { }
 
