@@ -14,7 +14,7 @@ export class InterceptorService {
     const token = this.tokenService.getToken();
     if(token != null){
       intReq = req.clone({
-        headers: req.headers.set('authorization', 'Bearer' + token)
+        headers: req.headers.set('Authorization', 'Bearer' + token)
       });
     }
     return next.handle(intReq);
