@@ -1,27 +1,28 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { persona } from '../model/persona.model';
+import { environment } from 'src/environments/environment';
+import { Persona } from '../model/persona';
 
 @Injectable({
   providedIn: 'root'
 })
 
 export class PersonaService {
-  URL = 'http://localhost:8080/personas/';
+  //URL = "http://localhost:8080/personas/";
+  URL = environment.EURL + '/personas/';
+  constructor(private httpClient : HttpClient) { }
 
-  constructor(private httpClient: HttpClient) { }
-
-  public lista(): Observable<persona[]>{
-    return this.httpClient.get<persona[]>(this.URL + 'lista');
+  public lista(): Observable<Persona[]>{
+    return this.httpClient.get<Persona[]>(this.URL + 'lista');
   }
 
-  public detail(id: number): Observable<persona>{
-    return this.httpClient.get<persona>(this.URL + `detail/${id}`);
+  public detail(id: number): Observable<Persona>{
+    return this.httpClient.get<Persona>(this.URL + `detail/${id}`);
   }
 
-  public update(id: number, Persona: persona): Observable<any>{
-    return this.httpClient.put<any>(this.URL + `update/${id}`, Persona);
+  public update(id: number, persona: Persona): Observable<any>{
+    return this.httpClient.put<any>(this.URL + `update/${id}`, persona);
   }
 
 }
